updated test and fixed part of #21
I completely rewrote the tests in a pythonic way. Now running
these tests is as easy as python -m unittest discover in test/

I'm planning to add more tests and especially unittest for
command line parsing which has several issues (-mllvm flags)
for instance.

The only test I did not move is :

`mystery:
	otool -X -s __LLVM __llvm_bc main > main.otool
	xxd -r  main.otool
	xxd -r  main.otool main.xxd`

as I could not understand what it does and could not find otool
on my ubuntu system. (any help there ?)

The fix for as is because it is not in the top-level directory.