Add support for FORTRAN code via the brand-new Flang front-end.
As part of the process, I've had to refactor the code to support more than just
a dual C/C++ codebase. This probably isn't the best means of doing so, but it is
sufficient for getting a working version of fortran compiling. Support for
Dragonegg+gfortran is added as well, but it's completely untested.

The code here is tested with the SPEC CPU2006 benchmark.